pyproject.toml
[tool.poetry]
name = "slack-message-remover"
version = "0.1.0"
description = ""
authors = ["donghwan <donghwan@twinny.co.kr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.23.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
